bunch of old/unused files from previous attempts
these files do not affect the currently hosted website in any way, just a bunch of failed attempts and test files used to get to the final result. You can see some progress here. BUT please note that these files will not run since they either require some libraries which can be installed but also csv/xlsx data files read were read from online and not locally so it would not give any outputs if ran, this is more just to view the code that led to this point rather than running the old files.